Lua: Fix reconnect issues due to GC client and will options strings
On ARMv7 I was experiencing following errors during reconnects:

 New connection from 192.168.1.20 on port 1883.
 Malformed UTF-8 in will topic string from remove_value, disconnecting.

 New connection from 192.168.1.20 on port 1883.
 Socket error on client <unknown>, disconnecting.

 Malformed UTF-8 in will topic string from `w, disconnecting.
 Socket error on client <unknown>, disconnecting.

This issues are caused by strings client_id and will topic/payloads
strings which are passed from Lua and could be eventually garbage
collected. To fix this issue we can't use strings as passed from Lua but
we've to use copies of the strings.

Signed-off-by: Petr Štetiar <ynezz@true.cz>